new_approach/adk-example/requirements.txt
google-adk
# Add any other dependencies your agent needs
